+++ src/main/java/bumblebee/xchangepass/domain/user/service/UserRegisterService.java
@@ -3,7 +3,8 @@ package bumblebee.xchangepass.domain.user.service;
 import bumblebee.xchangepass.domain.user.dto.request.UserRegisterRequest;
 import bumblebee.xchangepass.domain.user.entity.User;
 import bumblebee.xchangepass.domain.user.repository.UserRepository;
-import bumblebee.xchangepass.domain.wallet.wallet.service.WalletService;
+import bumblebee.xchangepass.domain.wallet.wallet.service.impl.WalletServiceImpl;
+import bumblebee.xchangepass.domain.wallet.wallet.service.impl.lock.PessimisticLockWalletService;
 import bumblebee.xchangepass.global.error.ErrorCode;
 import bumblebee.xchangepass.global.exception.CommonException;
 import bumblebee.xchangepass.global.util.DuplicateKeyExceptionHandler;
@@ -19,7 +20,7 @@ public class UserRegisterService {
 
     private final UserRepository userRepository;
     private final NicknameGenerator nicknameGenerator;
-    private final WalletService walletService;
+    private final WalletServiceImpl walletService;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
     /**
